refactor: renaming functions and consolidating files
